Move OperatorInputStats to top level
